moved everything to FolderHome.js
